requirements.txt
# Core AI/LLM SDKs
anthropic>=0.40.0
openai>=1.50.0

# Databricks
databricks-sdk>=0.36.0
databricks-sql-connector>=3.0.0

# LangGraph / LangChain (for OpenClaw orchestration)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0

# Vector Search / Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Video Animation
matplotlib>=3.8.0
Pillow>=10.0.0
imageio>=2.31.0
moviepy>=1.0.3

# Canva / Design
requests>=2.31.0
httpx>=0.27.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0

# Async support
anyio>=4.0.0
aiohttp>=3.9.0

# Utilities
rich>=13.7.0
click>=8.1.7
tenacity>=8.2.3
structlog>=24.1.0
